Graceful failure when posts/pages non-existent
Previously, the application would crash if either the configured posts
or pages directory did not exist. Given that the use of both pages and
posts should be optional, this was a bug.

Under these circumstances, a warning is now outputted to stderr, but the
app proceeds to boot otherwise normally.

Bumped to version 2.0.1.